Add user settings for folding any file. Fold on file save
Two settings are added, namly "base64fold_any_file" and
"base64fold_file_extensions", which can be set in the usual place
(Packages/User/Preferences.sublime-settings). See the descriptions in
the settings file for details.

To allow folding in any files, some adjustments are made to the regular
expression. Also, if the Base64 code length was not valid (not being a
multiple of 4), it won't be folded.

.htm and .html are added to the default file extensions list.

Another new feature is that Base64 code will be folded while the file is
being saved, not just after the file is openned.